Fix: Reuse fixtures
