+++ test/Fixture/Definition/Definitions/Acceptable/PersonDefinition.php
@@ -11,18 +11,19 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\DoesNotImplementInterface;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\Acceptable;
 
+use Ergebnis\FactoryBot\Definition\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 
 /**
- * Is not acceptable as it does not implement the DefinitionInterface.
+ * Is acceptable as it implements the interface.
  */
-final class UserDefinition
+final class PersonDefinition implements Definition
 {
     public function accept(FixtureFactory $factory): void
     {
-        $factory->defineEntity(Fixture\FixtureFactory\Entity\User::class);
+        $factory->defineEntity(Fixture\FixtureFactory\Entity\Person::class);
     }
 }
