+++ test/Fixture/Definition/Definitions/DoesNotImplementInterface/PersonDefinition.php
@@ -11,19 +11,18 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\IsAbstract;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\DoesNotImplementInterface;
 
-use Ergebnis\FactoryBot\Definition\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 
 /**
- * Is not acceptable as it is abstract.
+ * Is not acceptable as it does not implement the DefinitionInterface.
  */
-abstract class UserDefinition implements Definition
+final class PersonDefinition
 {
-    final public function accept(FixtureFactory $factory): void
+    public function accept(FixtureFactory $factory): void
     {
-        $factory->defineEntity(Fixture\FixtureFactory\Entity\User::class);
+        $factory->defineEntity(Fixture\FixtureFactory\Entity\Person::class);
     }
 }
