+++ test/Fixture/Definition/Definitions/FakerAware/PersonDefinition.php
@@ -11,16 +11,16 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ExtendsAbstractDefinition;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\FakerAware;
 
-use Ergebnis\FactoryBot\Definition\AbstractDefinition;
+use Ergebnis\FactoryBot\Definition\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 
-final class UserDefinition extends AbstractDefinition
+final class PersonDefinition implements Definition
 {
     public function accept(FixtureFactory $factory): void
     {
-        $factory->defineEntity(Fixture\FixtureFactory\Entity\User::class);
+        $factory->defineEntity(Fixture\FixtureFactory\Entity\Spaceship::class);
     }
 }
